Add original_file_path field to WooRequestController and make it nullable in the database
This commit introduces the 'original_file_path' field in the WooRequestController's storeManual method, initializing it to null upon request submission. Additionally, a new migration is added to modify the 'original_file_path' column in the 'woo_requests' table, allowing it to be nullable, which enhances the flexibility of the data model.